backend/requirements.txt
fastapi==0.129.0
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pandas==2.3.2
openpyxl==3.1.5
python-multipart==0.0.20
PyJWT==2.10.1
mcp==1.13.1
python-dotenv==1.1.1
email-validator==2.2.0
google-genai==1.64.0
google-adk==1.25.1
psycopg[binary]==3.3.3
cryptography==46.0.1
